fix warnings and Add validator classes to a separated file in signup.dart
